[dtrans] cmplrs-50816: AOS-to-SOA heuristic override should only apply to structures that meet safety criteria.
This fixes a bug introduced in the change for
https://git-amr-2.devtools.intel.com/gerrit/#/c/130101/ that allowed the command
line heuristic override flag used for internal testing to force the
transformation on types that did not meet the safety criteria. This requires
some changes to one of the test cases to mark some portions of that test as
FIXME because the analysis does not handle marking the type as safe in those
situations yet.

Change-Id: I6f92acac01f24ec14534c3cd9126be952b33870b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/132276
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Old/split llvm commit ID: fb46e6cf672eb156a57c2025b0161f2345a8b8c3