[dtrans] Update byte-flattened GEP instructions during DeleteField
This change adds support in the DeleteField optimization for updating
GetElementPtr instructions that use an i8* value and an offset to access a
field in a structure that is being optimized.

In the case of normal GEP instructions indices for non-deleted fields
couldn't be updated until after the instruction was replaced with the
new type, but because byte flattened GEPs use an i8* base pointer they
can be handled completely before the cloning.

CMPLRS-45540

Change-Id: I0f6a8a9f240689397a59311f747c13fcc6588976
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/131359
Tested-by: xmain_checkin_sanity <lab_icls@sceler38.sc.intel.com>
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: 4539ce2615370ebc3fdbf6d521cc4b10149197d7