Make AP-Int act more like an integer type when tested.
Reported in CMPLRS-50903, I have another case where a test for Integral
Type (or EnumType) causes an assert to fail.  I resisted making this
change earlier in the process, since it threw a bunch of other stuff
off, but it seems like this is a good time to do so.

It DID mess up handling of boolean/integer conversions due to an
ordering issue when checking, but it was easy enough to fix.  Not yet
doing 'hasXXIntegerRepresentation' since there is more logic to that.

Change-Id: Ia79494a57cedd1cb3b2834d21ecc8495f24c0daf
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/133304
Tested-by: xmain_checkin <lab_iclt@sceler38.sc.intel.com>
Tested-by: Felipe de Azevedo Piovezan <felipe.de.azevedo.piovezan@intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Reviewed-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Old/split clang commit ID: 45e862a703d2635a07052b2d28278561a98f386b