Fix getAllOnesBlockTypeRec
Structure of a divergent loop should obey the following scheme:

               entry
                 .         .
  all-ones -> all-ones      .
               . .           .
            test all zeroes  .
            .    .          .
           .   entry2      <- entry
           .     .
           .  original  <- original
            .    .   .
              . exit

Previously SINGLE_BLOCK_LOOP_ENTRY_TO_ORIGINAL was supposed to be handled in
self-loop handling. But it's not necessary for an all ones block to be a
self-loop and for a branch:

        SINGLE_BLOCK_LOOP_ENTRY
                  .
                  .
  SINGLE_BLOCK_LOOP_ENTRY_TO_ORIGINAL
                  .
                  .
      SINGLE_BLOCK_LOOP_ORIGINAL   <-- That is missed!
                  .
                  .
        SINGLE_BLOCK_LOOP_EXIT

a transition from entry to original could be missed.

So here instead of using an assertion I an add extra check for
SINGLE_BLOCK_LOOP_ENTRY_TO_ORIGINAL successor to restore proper
heritage.

Also, some odd comments were removed from the code.

Change-Id: I510ad2d663d5af377be1e2e745cc4b125de5b913
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/134048
Tested-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>