[NFC] [dtrans] Move base class storage of Function to CallInfo mapping from function scope to class scope.
This change moves the mapping of Function to CallInfo objects used to keep
track of which CallInfo objects need to be updated as each function is
cloned or remapped from being within the transformation function to being
within the base class. This is to prepare for the derived classes needing to
update or delete CallInfo objects when calls to the memfunc routines get
eliminated by the derived class. Interface routines that can be invoked
by the derived classes to inform the base class of changes will be added
in a subsequent changeset.

There are no functionality changes in this patch.

Change-Id: Ia003816786d3a3469ac320d79d69684fc3f9c380
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/134174
Reviewed-by: Chupina, Olga <olga.chupina@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Ansari, Zia <zia.ansari@intel.com>
Old/split llvm commit ID: e8c4757f150da5acd7fb9927c4c73076d8f320bf