[dtrans] Update initializers for global fixed arrays
This change implements code to update the initializer for global
variables that are a fixed size array of a structure from which fields
have been deleted.

CMPLRS-50969

While developing this patch, I realized that GEP instructions will also
not be updated correctly if the GEP is accessing fields through a fixed
array object. I did not fix that problem in this change set, but I did
create test cases that exhibit the problem and filed a JIRA. It is not
necessary to fix this problem for the 557.xz benchmark, which is the
benchmark motivating CMPLRS-50969.

Change-Id: I678a1d8263f3aa2c1138e356b75f324b5d09cd02
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/134354
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: 55cb6277b6371bdb44c100e12d6061d41fb376e6