[dtrans] AOS-to-SOA support for byte-flattened GEPS
Transforms byte-flattened GEPS into a sequence of instructions
that finds the address of the field in the structure of arrays.

Also updates the handling of bitcasts that are either from or to
pointers of types being transformed.

Moves instruction deletion until the end of processFunction to
avoid potential for caching a pointer to an instruction that has
been deleted.

Change-Id: I4f087e438a024cb6e5828c34f339221e70325cad
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/134945
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@sceler38.sc.intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Ansari, Zia <zia.ansari@intel.com>
Old/split llvm commit ID: 570f5bc12c7ce145516d4bbd9f8623341b34392b