[dtrans] Fixes for base class conversions of GlobalValue objects
This change addresses two problems in the DTrans optimization base class:
1. Global variable initialization needs to handle all global variables, not
just variables whose type change. This is because the initializer constant
for any variable may refer to a variable or function that is being replaced.
2. Aliases need to be updated when the aliasee is going to be replaced.

This also changes the map of original functions to clone functions to use
a MapVector, so that when the items are walked for removal, it happens
in a consistent order. The order does not matter for correctness, but
can impact debugging issues. This can be changed back to a DenseMap in
the future.

Change-Id: Ia94ed276a78d5990e3f080dabfa2abfdd9039654
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/135413
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Old/split llvm commit ID: d45848c060aaac8e245297ae1dc840644c51a6ed