Fix unconditional execution of scalar version of kernels with barrier.
Due to an overlook in StripIntelIP wrapper of a vectorized kernel was
not spared meaning that link between scalar wrapper and its vectorized
counterpart was lost. 'vectorized_kernel' metadata of a scalar wrapper
was null and vectorized version was never executed.

LIT test illustrates the expected layout of wrappers, scalar and
vectorized kernels.

Change-Id: I34f4dabadd7afb198f370786443d1deefe3c6afb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/136221
Tested-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Bessonova, Kristina <kristina.bessonova@intel.com>