Remove transFlags in SPIRVToLLVM
transFlags sets the "exact" flag in LLVM-IR for AShr and LShr
instructions. This is in fact incorrect as this means that LLVM will do some
optimisations that are potentially illegal.

Setting the exact flag says that there are no set bits in the bits being shifted
off the end during the shift - and thus shift combining is safe to do.

In a situation like this ((X >> 5) << 2) it will be transformed into (X >> 3)
which is incorrect as the 2 LSBs may contain set bits.

This change is ported from AMDVLK

Change-Id: I9819299656e5e580a31fc65457c8934cf9f32fad
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/168016
Reviewed-by: Saito, Hideki <hideki.saito@intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: Sotkin, Alexey <alexey.sotkin@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Old/split llvm commit ID: b9596a5c696fff62c5398286790e08fb8186bccb