CMPLRS-51285: Fix binary ops with bool and AP-Int with bitsize<8
Previously, conversions were done assuming that all of the operands that
made it through the Sema checking would be integers and did conversions
based on the bitsize of the variable.  Unfortunately, bool has a bitsize
of 8, so any ap-ints smaller than that would not be properly converted.

This patch adds special handling for the boolean type so that
conversions are properly applied.

Change-Id: I89cc01403eb270b2a67168931ec4b42c95c750ca
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/137680
Reviewed-by: Hoffner, Tommy <tommy.hoffner@intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Tested-by: xmain_checkin <lab_iclt@sceler38.sc.intel.com>
Reviewed-by: Ansari, Zia <zia.ansari@intel.com>
Old/split clang commit ID: cd5be4ea094f77df25a78b0ab0ead6e58960206c