CMPLRS-50651 - Add NORMALIZED.UB clause for the upper bound.
For 'omp simd', Sema does not build the loop expressions with a bounds
variable.  So to make simd loops look like the others we have to create
a ub variable and also create a condition expression that uses it.  These
expressions are only used in the simd case.

Change-Id: Ie9f64d9bb629fc7946a58f942718cecaa5d97db0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/137970
Tested-by: xmain_checkin <lab_iclt@sceler38.sc.intel.com>
Reviewed-by: Yu, Jennifer <jennifer.yu@intel.com>
Reviewed-by: Saito, Hideki <hideki.saito@intel.com>
Old/split clang commit ID: 3998e021504a4770b1762414042538c5333026a4