Correct Integer-Rank behavior for AP-Ints
As reported via email by Vince, mixed-type ternary operators were not
being properly handled when the operands were arbitrary precision
integers.  This patch adds them as arithmetic types, so that they can be
properly considered, then gives it an integer rank.

As with the rest of the integer ranks, make the ap-int the shifted
version of its size.  We add 7 so that the ap-ints will have a higher
rank than any other int-type (with the exception of int128 due to
implementation constraints, but I believe that is acceptable).

Change-Id: I03c4aa6902e25d143955140fbb3cc9f64dd125e4
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/138295
Tested-by: xmain_checkin <lab_iclt@sceler38.sc.intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Reviewed-by: Ansari, Zia <zia.ansari@intel.com>
Old/split clang commit ID: 276c27b92c83ce30a4342125748d8eae270ddee7