CMPLRS-51394: Fix boolean handling in HLS stream calls.
Boolean types are special in that they are represented as both an i1 and
an i8 in clang/llvm depending on the situation.  All pointers are
supposed to be i8*, but our intrinsics want to be able to handle i1* (in
order to differentiate between bool and char pointers).

To accomplish this, we just need to make sure to bitcast the results on
the way in/out of this function. MemMaster calls already did (since it
used a single return, it used EmitCall), so this just does the same as
that one.

Change-Id: I3e82a13c713217472b23633704b9ddbe4e77ea6e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/138616
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: Keane, Erich <erich.keane@intel.com>
Reviewed-by: Ansari, Zia <zia.ansari@intel.com>
Old/split clang commit ID: 77914db7f30bcf59421b2746f16c18628c93c401