[dtrans] Handle zero in isValueMultipleOfSize
CMPLRS-50652

Under various circumstances the DataLayout object will report that a
type has an allocation size of zero. Common cases include structures
with no elements and zero-sized arrays. This patch modifies
dtrans::isValueMultipleOfSize to always return false if a size of zero
is passed in.

We may eventually want to handle the zero-size array differently since
that is actually used to represent an array of unknown length, but that
would have to be done where the instruction using the array is handled
and not in this function.

Change-Id: I150c14df703196e7d835292fb48947f525a0272b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/139419
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@sceler38.sc.intel.com>
Reviewed-by: Saito, Hideki <hideki.saito@intel.com>
Old/split llvm commit ID: 4ad174c258e697af57c4f129236a52c0be7287a7