Fix debug info while cloning function in CLWGLoopCreator
If a vector function was inlined in CLWGLoopCreator pass into an original
function, the first one is supposed to be removed, but its DISubprogram is left
in a module. It happens because of the DISubprogram has users - DILocations of
instructions inlined from 'old' vector function. These DILocations refer to
the DISubprogram as a scope. It leads to a crash in the CodeGen.

To let 'old' DISubprogram die we do the follow things:
- drop DISubprogram from the vector function to avoid duplicating it
while cloning,
- disallow CloneFunctionInto() doing module level changes,
- change a DILocation scope of all inlined instructions to DISubprogram
of original function.

TODO: It looks more logical to use llvm::InlineFunction instead of
manual inlining. It assumes llvm::InlineFunction is able to handle all
such issues with debug information.

Change-Id: I9aa8a395ea3fd0689b912fbf5b9c479cb81849bb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/139942
Tested-by: Nazarov, Mikhail V <mikhail.v.nazarov@intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>