[VPlan] Set master VPInstruction in UB decomposition
This is a fix for the decomposition of the upper-bound of the HLLoop.
After the decomposition, no master VPInstruction was set for the group
of decomposed VPInstructions. The fix sets the last VPInstruction of the
group as master VPInstruction and attaches the HLLoop as underlying IR.

The bug was exposed by the following changeset not committed yet:
https://git-amr-2.devtools.intel.com/gerrit/#/c/137459/
Mostly all the VPlan HIR test hit an assert without this fix.

The fix changes a property of a VPInstruction which is not printed or
impact the generated code so I'm not sure how to create an specific
regression test. The aforementioned failures should suffice.

Change-Id: Ice34410fc2c8ef04e9e58a345f577d5b87b40734
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/144833
Reviewed-by: Caballero, Diego <diego.caballero@intel.com>
Tested-by: Caballero, Diego <diego.caballero@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 09b42237c28e1b1628092be1ffc3ee177117b870