cmplrs-51726: Avoid Klocwork flagged potential for nullptr dereference
This change adds a conditional test to verify a pointer is non-null before
dereferencing it. This is a preventative fix, because the pointer is not
expected to be nullptr with the current implementation of the community
[post]dominator tree implementations, but the interface defines that
a call may return nullptr for an unreachable block.

Change-Id: I6a0bae5e91bc5f495d518b31e04855c9336da3a1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/141982
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@sceler38.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@sceler38.sc.intel.com>
Reviewed-by: Saito, Hideki <hideki.saito@intel.com>
Old/split llvm commit ID: 0eb61df64ca2bb9ac62c39262a5c48f41fc20861