Document dataflow conversion better.
This patch has one functional change that fixes CMPLRS-51102. All of the other
changes are for better documentation or better debugging:

* Documentation comments are added to CSACvtCFDFPass.h for functions related to
  pick tree creation and patching. This also moves the functions together to
  add them to a single dataflow group. Parameter names are modified in the
  declarations but remain untouched in the definitions.

* Extra output is added to -debug-only=csa-cvt-cf-df-pass to describe what is
  going on with pick tree creation. This requires one change to functions so
  that the debugging output can correctly identify which edge is causing the
  pick tree entry to be installed.

* The debug dump also displays the dataflow output after switch generation, and
  again after pick generation (but before LIC group logic or basic block
  deletion). This reduces the need to catch stuff in the debugger.

* Extra comments are added to describe some of the cases.

* Sphinx documentation is added to describe the text.

The sole functional change fixes CMPLRS-51102, and the gory details of why that
change is needed is described over there.

Documentation for generating the pick tree (but not patching it) should be
complete. Everything else is known to be incomplete, and I intend to fix the
documentation, but this patch should provide a good start for improving the
documentation of this code.

Change-Id: I8413dec17e9951001e5e928d65b0a932be502b30
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/141329
Reviewed-by: Cranmer, Joshua <joshua.cranmer@intel.com>
Tested-by: Cranmer, Joshua <joshua.cranmer@intel.com>
Old/split llvm commit ID: 0ccf083502589f7b7f308a5720a9611bd0464929