Avoid LIC not empty on exit warning messages.
In a pick-switch loop, the input control value for the pick is a shifted value
of the control value for the exiting switch. This means that when such a loop
exits, the initial value is sent to the pick for the first iteration of the
next loop invocation. If this input to the pick is constant, then it causes
part of the loop to start executing.

Change-Id: I8c69f1f421733e80fa914b529f38000257b8831f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/143334
Reviewed-by: Woodworth, Daniel <daniel.woodworth@intel.com>
Tested-by: Cranmer, Joshua <joshua.cranmer@intel.com>
Old/split llvm commit ID: fa4f7e598aff3b01e1c5f718d8d80f72e09c2f21