[dtrans] Recognize ptr-to-ptr element zero bitcasts
When a pointer-to-pointer to a structure is bitcast as a
pointer-to-pointer to element zero of that structure, we need to
recognize this and not treat it as a bad cast. This is distinct from a
standard element zero access in that the resulting pointer is not itself
a pointer to the element.

This pattern looks like it will be common in C++ where element zero of a
class is the base class.

This patch resolves some bad bitcasts and mismatched element access
safety conditions for the 523.xalanc benchmark.

At this point I'm leaving open a potential issue I noticed while
developing this patch where we would not detect a read of an element
zero pointer obtained in this way. I added a test for that case and have
field a JIRA report. I have not seen this pattern in any program or
benchmark.

CMPLRS-51516

Change-Id: Ia51bb43a46f366b45017d030f4f40d655b69f029
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/140973
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: 4d4c68e0e9c04da90ce8d7193112333a63d4abfe