Add a test to reveal incorrectly updated BB frequencies during critical edges splitting.
Summary:
The SplitIndirectBrCriticalEdges function introduces an issue appearing in
updating of BB frequencies in presence of multi-edges from "direct"
predecessors. Single user of the code now is PGO instrumentation pass,
but it doesn't preserve any analysis to check incorrect BB frequencies
after the pass. Due to that the issue can be exposed via dumping of
debug information during PGO instrumentation pass.

Branch weights defined in the test are chosen to result in 0 BB frequency
assigned to a block that left as empty BB on a critical edge split (sw.bb)
in presence of the issue. It will result further in 0 weight assigned to
a branch from this BB during MST creation, which is reflected in debug dump.

Change-Id: Id7b97b53af48e147eef469c50bff936b51324dee
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/144748
Tested-by: Shlykov, Arkady <arkady.shlykov@intel.com>
Reviewed-by: Shlykov, Arkady <arkady.shlykov@intel.com>
Old/split llvm commit ID: a1e9f188e54f712467e5733baede44a8a705c247