[VPlan] Introduce 1st-iteration peeling for multi-exit loops
This patch introduces the starting point of a peeling utility for
HLLoops in HIR. Currently, it peels the first iteration of an HLLoop
but it should be easily extendible to N-iteration peeling.

The utility is used by VPlan HIR CG to peel the first iteration of
multi-exit loops. This is a workaround whereas we don't have the
proper peel/remainder modeling in VPlan. This peeling is necessary
for 557.xz multi-exit loops.

The current approach uses a flag to enable the 1st iteration peeling
of vectorized multi-exit loops. This flag will be replaced with the
proper idiom recognition detection once it's ready.

Change-Id: I20edf0a3845fe236577c6304497407cd91283d41
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/145263
Reviewed-by: Caballero, Diego <diego.caballero@intel.com>
Tested-by: Caballero, Diego <diego.caballero@intel.com>
Old/split llvm commit ID: ab04a3b394119f5b049ade5e3a63657d5352368c