[dtrans] Handle i8 element zero store
This patch fixes a case where a store of an i8 value to the first
element of a structure was being misidentified as an unsafe pointer store.

We have special code to treat a getelementptr with all zero indices as
if it were a bitcast to the parent structure type (because LLVM's
instcombine converts such a bitcast to a GEP), and so we were not
tolerating the case where a i8 value is stored to the i8* pointer
obtained this way.

CMPLRS-51173

Change-Id: I314472251b1f794939fb8331309e9b53afdac9b9
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/141227
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Reviewed-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Old/split llvm commit ID: 1c4755f25b56c714949a5e4a9c65cbee222b8be3