[dtrans] Add check that dtrans analysis completely analyzed module into dtrans opt base class
The base class relies on the information in the DTransAnalysis class for
computing dependent types. This change makes sure that the analysis did
not take one of the early exit conditions that skips the analysis because
this could lead to incorrect IR.

Also fixes DTransAnalysisInfo::useDTransAnalysis to behave like the
comment described it as behaving.

Change-Id: I34f2925c5edded3173ab6f803fb3d4d668aec952
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/146710
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Old/split llvm commit ID: be488f24f561d03bd7b580387d43d591bf0126b8