[NFC] Extend NORMALIZED.IV/UB to support a list of IVs/UBs instead of just 1
Clang FE sends the IV and UB of an OMP loop to the BE via clauses: 
   "QUAL.OMP.NORMALIZED.IV"(i64* %.omp.iv)
   "QUAL.OMP.NORMALIZED.UB"(i64* %.omp.ub)

To support offloading OMP PARALLEL FOR loop nests to OpenCL, we need
the IV and UB of every loop in the loop nest. This change extends the
representation to allow a list of IVs and UBs in the corresponding
clauses. Eg, for a loop nest of depth 2, we'd have:
   "QUAL.OMP.NORMALIZED.IV"(i64* %.omp.iv0, i64* %.omp.iv1)
   "QUAL.OMP.NORMALIZED.UB"(i64* %.omp.ub0, i64* %.omp.ub1)

Change-Id: I725277d81fb04394a76a9d5ee341150a8834e209
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/147092
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Lin, Jin <jin.lin@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: c6793d1a0ea44e36161d5b1f8dfe2c83ac8fa8de