[LoadCoalescing][CMPLRLLVM-1047 ] Changes for implementing load-coalescing.
This pass scans for vector loads that are accessing consecutive
memory locations and attempts to replace them with a wider loads.
For example, if there are two 4-wide loads from A[i+3:i] and
A[i+7:i+4], it  will try to replace them with an 8-wide A[i+7:i],
if this is legal.

This pass is supposed to work in conjunction with SLP Vectorization

Change-Id: Id0e53a78c565cf53fd9d0b8886fb4eb8cb8034ca
Jira-Id:CMPLRS-52235
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/151054
Reviewed-by: Brevnov, Evgueni V <evgueni.v.brevnov@intel.com>
Tested-by: Brevnov, Evgueni V <evgueni.v.brevnov@intel.com>
Old/split llvm commit ID: a879ec9a831b6c804879ad630bad2977e39a7e50