[dtrans] Analyze compatible type use in ResolveTypes (NFC)
This change adds code to analyze the uses of types that are not
equivalent but differ only by the types of pointers in one or more
fields. This analysis is gathering information that will later be used
to decide whether or not it will be profitable to remap these types.

This change should not produce any functional change in the results of
the pass.

CMPLRS-51516

Change-Id: If90b0158a1b50c98243ce28558c467b09e5d8b9f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/144332
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chupina, Olga <olga.chupina@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 2aebdff836f24726c86876e7d89e4cc59ee04505