[dtrans] Add an option to disable the ResolveTypes pass
The 557.xz benchmark was getting a compile failure under some
circumstances due to a problem in the ResolveTypes pass. This problem
will be fixed, but it seemed useful to have an option to disable the
pass in case issues like this arise in the future.

Change-Id: I6b1afc7f32141056cf0b2b5c2e1c6c08cdf4cde8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/148377
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 144800bfd8eda3ff6473d936460d03d3416beab7