[dtrans] Fix problems with ResolveTypes
This addresses a problem that was observed when running the 557.xz
benchmark where ResolveTypes was remapping a type without also remapping
the nested type.

The problem looked like this:

%union.anon.0 = { %struct.anon.34 }
%union.anon.0.35 = { %struct.anon.26 }
%struct.anon.26 = { [8 x i32], i64 }
%struct.anon.34 = { [8 x i32], i64 }

The ResolveTypes pass was remapping %union.anon.0.35 to %union.anon.0
but it was not remapping %struct.anon.26 to %struct.anon.34 so existing
GEP instructions failed verification due to a type mismatch.

Because of the change to rename structures when the type with the base
name does not exist, it was necessary to trim multiple numeric suffixes
such as:

%struct.A.1.2.3

This actually matches what the LLVM IR linker attempts to do so I think
that is a safe solution.

Change-Id: I1f3bb689db2ca657e0dc964967a544f38b275faf
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/148552
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: 47a58b4318a5a3947d97154ba75acae6053d8175