[AddSubReassociation] Support for OR instruction.
ADD op1, op2 instruction may be replaced with OR op1, op2 if it is proven that they may not have common bits set. With this change we interpret such OR instruction as an ADD what effectivly makes it part of a tree.

Change-Id: I386a0de8a085d8063b842eb7f5528dafe251b7e2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/148580
Reviewed-by: Brevnov, Evgueni V <evgueni.v.brevnov@intel.com>
Tested-by: Brevnov, Evgueni V <evgueni.v.brevnov@intel.com>
Old/split llvm commit ID: 7c11733d3668b2e82ad7039265e7e2ed205180e1