[FPGA] Add ChannelsUsageAnalysis pass to print channels usage graph
This is pass is actually NOOP until 'VOLCANO_CHANNELS_USAGE_GRAPH' env
variable is not specified. Pass generates a text file which contains
graph of kernels connected with channels in .dot format.

Pass also can be used via oclopt -analyze -channels-usage-analysis.

Not all TODO's are resolved in this patch, but current functionality is
enough to cover most of the benchmarks.

Change-Id: I3eb8faf1f17b39af7cb1a05dc9e88dfacf7e4662
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/149046
Reviewed-by: Sachkov, Alexey <alexey.sachkov@intel.com>
Tested-by: Sachkov, Alexey <alexey.sachkov@intel.com>