Fix SROA crash when var.annotation is called with a GEP.
There is a mismatch between what SROA sees as promotable and what
PromoteMemoryToRegister asserts to be promotable. In particular,
SROA is ok with a GEP being used by a var.annotation whereas
PromoteMemoryToRegister is not.

This bug was likely introduced by:
https://git-amr-2.devtools.intel.com/gerrit/#/c/117529/

The fix is to simply change PromoteMemoryToRegister, which already
allows a bitcast to be used by a var.annotation (but not a GEP).

To reproduce the issue, the LIT contains lifetime call to the alloca.
SROA considers lifetime markers as uses of the alloca, whereas the same
is not true for annotation calls. If only annotation calls are left,
SROA will not try to do anything because it sees no uses of the alloca.

Change-Id: Id5169f7507256338b24ab2c43c620ccc48ab781c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/149074
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Begam, Rehana <rehana.begam@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 1637660b69fe96a118635430a42b2c004528d6a4