[dtrans] Modifications for DTrans AOS-to-SOA conversion of allocation sites to enable dependent structures to pass DTrans safety checks
This modifies the transformation of the allocation site to allow annotating
the instructions that store the pointers for the individual arrays being
partitioned out of the allocated the memory block with the type of array.
This is used to avoid the analysis from treating these as bad casting or unsafe
pointer stores when analyzing IR following the AOS-to-SOA transformation.

The DTransAnalysis is modified to now look for metadata annotations on
instructions for a type to associate with the instruction result.

Change-Id: Id22792d45416822b6879f564c8eaf52599ee020b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/149376
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 41283acb0221f87cc370dc46def4c1a7204e411c