[dtrans] Implement compatible type merging heuristic
This change implements a simple heuristic to determine whether or not
compatible types should be remapped.

If one type is seen to be converted to another in a function bitcast at
a call site, and the source type does not access any fields that
conflict with the types of the same fields in the destination type then
we would like to remap these types.

Depending on other remappings that have been previously performed, this
may not be possible.

Change-Id: Ibe2998e34839ec473e010ab12c399acc97e55cf6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/146142
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chupina, Olga <olga.chupina@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 449ddcf66ecd951f0286f798dc904d312bdbfe4a