[NFC]VPO: Add a field to Target WRN to support kernel parallelization
For openCL-based offloading of

  #pragma omp target
    #pragma omp parallel for

when we process the "parallel for" we create a struct for ND-range info
that is needed later when processing the "target" construct.

This change adds an AllocaInst* field to the Target WRN to keep track of it.

Change-Id: I4e243f639bcd244bbaf6d4b6aea46018d2682c32
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/150994
Reviewed-by: Lin, Jin <jin.lin@intel.com>
Tested-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: d598480af4aa1f0be333c9a9e4213a88dedbdc76