[FPGA] Workaround issue with S10-optimized gzip when -g specified
Problem description:

LLVM spec doesn't see a difference between two or more structures having
same inners - they will have the same layout in the memory if there is no opaque
type inside. So if we are moving one IR module into another and both of them are
having structures with equal inners these structures are being mapped and the
second appeared is being substituted with the first one.

While struct pipe_t is a complex struct having a struct of 3 integers
inside if a user declared a struct of 3 integers as well it was leading
to a situation when this inner struct was substituted with a user's one
for pipe built-in prototype moving from a source module to a destination.
IRMover inside LLVM couldn't verify that a call inside a destination
module and the prototype were the same instances because of the types
mismatch and refuses to finish the built-in importing.

Pipe built-ins were refactored to use void* as pipe arguments: actual
cast to struct __pipe_t happens inside built-ins. This change removes
%struct.__pipe_t from a destination module and the original issue
disappears.

Change-Id: I9cc3c934bb2dd931d7ae1dcb3914abd19e9b8fd5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/151097
Reviewed-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>
Tested-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>
Reviewed-by: Savonichev, Andrew <andrew.savonichev@intel.com>