Refactor dataflow conversion to improve loop handling and dynamic predication.
This ends up grouping a few not-entirely-unrelated changes into one patch. The
main brunt of the patch is a cleaner method of dynamic predication; this method
reuses predmerge instructions to compute both the pick tree and the block
predicates. Testing indicates it should work on every CFG at this point that
can be thrown at dataflow conversion (i.e., single return, no infinite loops,
and no irreducible loops).

The second portion of this patch moves loop handling to a slightly different
phase. The long-term hope of this is that when we need to implement static loop
pipelining/unroll-and-jam, which involves modifying the condition of the loop
header pick in a similar manner to inner loop pipelining, we can unify that
handling much like inner loop pipelining does. I had hoped to make the loop PHI
logic be unified between structured PHI and dynamic predication, but the
resulting code was more complex than necessary.

Structured loop phi handling was also changed here. The loop header PHIs are
replaced with a pick between starting the loop and continuing the loop header
execution. This change allows things such as multiple-latch loops to be
supported properly with structured PHI generation. More complex conditions such
as while (a && b) are currently desupported, but the old code was already a bit
too optimistic about its ability to handle such cases.

There are few other small fixes for bugs that were found with an aggressive
approach to using dynamic predication. Structurize CFG should no longer be
necessary, but there is an impressively complex miscompile in sequence
optimization that is currently hidden when it runs, so it is still enabled by
default in this patch.

Change-Id: I8dfce80a5d5e2c10ac928e4d6320d87d9fb331ec
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/150355
Reviewed-by: Woodworth, Daniel <daniel.woodworth@intel.com>
Tested-by: Cranmer, Joshua <joshua.cranmer@intel.com>
Old/split llvm commit ID: 912de4b1fecbfe6f83f193a36454fd6fd9c3d641