[dtrans] Add analysis of ConstExpr uses of globals
When a global variable is used in a constant expression, the expression
appears directly as an operand in any instruction that needs it. As a
result the dtrans instruction visitor was not visiting these
expressions. The local pointer analyzer handles operands differently and
so it was already analyzing these cases, but any safety condition that
occurred as a direct result of the constant expression was missed.

I am removing some test cases that involved chained bitcast operand
because it turns out that these are compressed into a single expression
for use in the instruction before dtrans is called even when dtrans is
invoked directly from opt. The spurious bitcasts were still present in
the IR but not attached to any instruction and so they didn't have the
effect the test intended.

This patch also fixes a problem where byte-flattened GEP analysis was
not handling the possibility of an array of structures. It was necessary
to fix this problem in order to keep aostosoa-safety-deps02-index32.ll
as a valid and passing test.

This patch fixes CMPLRLLVM-7024.

Change-Id: I74236b0596228b90677acdf3274b9a2d8f5c0e69
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/152028
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: 263992bd59c4becfd1490bc085e76f710614492c