[NFC] [dtrans] Separate routines used for DTrans annotations to a new file
This change refactors the new utility functions that were added in the
commit for annotating the AOS-to-SOA memory allocation block to a
new utility class. This is in preparation for this utility class
to provide some generic functionality that can be used for other
DTrans specific annotations, and provide a place that a cleanup pass
will be able to identify all the DTrans specific annotations that are
to be cleaned up.

Additionally, documentation is added regarding the annotations, along
with a draft of the future annotation interface routines.

Change-Id: Ia34383e888f95ac71361086cd3acf7404fd26ee8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/152032
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Old/split llvm commit ID: 11148227a41fa1f5bf5428682fe86c61e0a7e916